refactor:Cambio de nombre de AddEditHeadquarter a AddEditHeadquarteCard
